update order filter, refund tests to merchant
