+++ tests/e2e/utils/src/flows/shopper.js
@@ -30,8 +30,6 @@ const gotoMyAccount = async () => {
 	} );
 };
 
-const WP_ADMIN_SINGLE_CPT_VIEW = ( postId ) => baseUrl + `wp-admin/post.php?post=${ postId }&action=edit`;
-
 const shopper = {
 	addToCart: async () => {
 		await Promise.all( [
@@ -181,23 +179,6 @@ const shopper = {
 			page.click( 'button[name="login"]' ),
 		] );
 	},
-
-	goToOrder: async ( orderId ) => {
-		await page.goto( WP_ADMIN_SINGLE_CPT_VIEW( orderId ), {
-			waitUntil: 'networkidle0',
-		} );
-	},
-
-	updateOrderStatus: async ( orderId, status ) => {
-		await page.goto( WP_ADMIN_SINGLE_CPT_VIEW( orderId ), {
-			waitUntil: 'networkidle0',
-		} );
-		await expect( page ).toSelect( '#order_status', status );
-		await page.waitFor( 2000 );
-		await expect( page ).toClick( 'button.save_order' );
-		await page.waitForSelector( '#message' );
-		await expect( page ).toMatchElement( '#message', { text: 'Order updated.' } );
-	}
 };
 
 module.exports = shopper;
